requirements.txt
flask~=3.1.1
pillow~=10.4.0
torch~=2.7.1
h5py~=3.14.0
transformers~=4.54.1
gradio~=5.40.0
huggingface-hub~=0.34.3
requests~=2.32.4
werkzeug~=3.1.3
flask-cors~=6.0.1
safetensors~=0.5.3
environs~=14.3.0
xformers~=0.0.31.post1